+++ src/quacc/__init__.py
@@ -7,7 +7,7 @@ from ase import Atoms
 from ase.io.jsonio import decode, encode
 
 from quacc.settings import QuaccSettings
-from quacc.util.wflows import flow, job, subflow
+from quacc.utils.wflows import flow, job, subflow
 
 __all__ = ["flow", "job", "subflow"]