+++ src/quacc/recipes/qchem/core.py
@@ -11,9 +11,9 @@ from quacc import job, SETTINGS
 from quacc.calculators.qchem import QChem
 from quacc.schemas.ase import summarize_opt_run, summarize_run
 from quacc.schemas.atoms import fetch_atoms
-from quacc.util.atoms import check_charge_and_spin
-from quacc.util.calc import run_ase_opt, run_calc
-from quacc.util.dicts import merge_dicts, remove_dict_empties
+from quacc.utils.atoms import check_charge_and_spin
+from quacc.utils.calc import run_ase_opt, run_calc
+from quacc.utils.dicts import merge_dicts, remove_dict_empties
 
 if TYPE_CHECKING:
     from ase import Atoms