+++ src/quacc/schemas/ase.py
@@ -13,10 +13,10 @@ from ase.vibrations.data import VibrationsData
 
 from quacc import SETTINGS
 from quacc.schemas.atoms import atoms_to_metadata
-from quacc.util.atoms import prep_next_run as prep_next_run_
-from quacc.util.db import results_to_db
-from quacc.util.dicts import clean_dict
-from quacc.util.files import get_uri
+from quacc.utils.atoms import prep_next_run as prep_next_run_
+from quacc.utils.db import results_to_db
+from quacc.utils.dicts import clean_dict
+from quacc.utils.files import get_uri
 
 if TYPE_CHECKING:
     from ase import Atoms