+++ CHANGELOG.md
@@ -80,7 +80,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 ### Added
 
 - Added support for Prefect.
-- Added `quacc.util.wflows` module.
+- Added `quacc.utils.wflows` module.
 
 ### Changed